tweak file name ans structure
